cheesecake

Used to disaggregate small area population estimates by age, sex, and other socio-demographic and socio-economic characteristics (e.g., ethnicity, religion, educational level, immigration status, etc).

It uses statistical models to predict population proportions and population totals across demographic groups. Primarily designed to help users in filling population data gaps across demographic groups due to outdated or incomplete census.

Usage

cheesecake(df, output_dir)

Arguments

df

A data frame object containing sample data (often partially observed) on age and sex groups population data as well as the estimated overall total counts per administrative unit.

output_dir

This is the directory with the name of the output folder where the disaggregated population proportions and population totals are automatically saved.

Value

A list of data frame objects of the output files including the disaggregated population proportions and population totals along with the corresponding measures of uncertainties (lower and upper bounds of 95-percent credible intervals) for each demographic characteristic. In addition, a file containing the model performance/model fit evaluation metrics is also produced.

Examples

data(toydata)
result <- cheesecake(df = toydata, output_dir = tempdir())
#> Directory C:\Users\ccn1r22\AppData\Local\Temp\RtmpA12aiY already exists.
#> [1] "(1) age_1 model is running"
#> [1] "(2) age_2 model is running"
#> [1] "(3) age_3 model is running"
#> [1] "(4) age_4 model is running"
#> [1] "(5) age_5 model is running"
#> [1] "(6) age_6 model is running"
#> [1] "(7) age_7 model is running"
#> [1] "(8) age_8 model is running"
#> [1] "(9) age_9 model is running"
#> [1] "(10) age_10 model is running"
#> [1] "(11) age_11 model is running"
#> [1] "(12) age_12 model is running"
#>           MAE       MAPE     RMSE corr
#> [1,] 1.002179 0.02859478 1.431711    1