Partially observed country-wide or sub-national demographically structured data by age and sex groups

Artificially generated toy data set that comes in a cross-sectional format where the unit of analysis is either administrative unit or enumeration area identifiable by an 'id' variable. 'It provides artificial information for 3,213 spatially distinct administrative units in which the individuals in the population are grouped into 12 mutually exclusive and exhaustive age groups. Each of the age groups was further grouped into 'male' and 'female' groups. The data contains the total population counts (total) for each spatial unit but also contains missing age and sex groups population counts. The model first predicts the population proportions of the mising data and then disaggregates the population totals using the predicted proportions to obtain the predicted population counts for the age and sex groups. Note that the same applies for other demographic groups such as marital status, race, etc. illustrate the use of the package.

Usage

data(toydata)

Format

An object of class "data.frame"

id

Administrative units ids (as numeric) for the administrative units of interest

x1,x2,x3

These are the samples of geospatial covariates (only required for the 'cheesecake' and the 'slices' functions). Note that these are the covariates identified to significantly predict population distribution among the demographic groups. The package allows the user to include any number of covariates in thier own datasets.

total

The total population counts being disaggregated and should correspond to the rowsum of the age groups or the sex groups population counts

age_1, ..., age_12

These correspond to the partially or fully observed number of people corresponding to each of the age groups. Note that we have used only 12 age groups here for illustration but the package can accomodate any number of age or sex or any demographic groups.

fage_1, ..., fage_12

These correspond to the partially or fully observed number of females corresponding to each of the age groups. Note that we have used only 12 age groups here for illustration but the package can accomodate any number of age or sex or any demographic groups.

mage_1, ..., mage_12

These correspond to the partially or fully observed number of males corresponding to each of the age groups. Note that we have used only 12 age groups here for illustration but the package can accomodate any number of age or sex or any demographic groups.

References

This data set was artificially created for the jollof package.

Examples


data(toydata)
names(toydata)
#>  [1] "id"      "x1"      "x2"      "x3"      "total"   "age_1"   "age_2"  
#>  [8] "age_3"   "age_4"   "age_5"   "age_6"   "age_7"   "age_8"   "age_9"  
#> [15] "age_10"  "age_11"  "age_12"  "fage_1"  "fage_2"  "fage_3"  "fage_4" 
#> [22] "fage_5"  "fage_6"  "fage_7"  "fage_8"  "fage_9"  "fage_10" "fage_11"
#> [29] "fage_12" "mage_1"  "mage_2"  "mage_3"  "mage_4"  "mage_5"  "mage_6" 
#> [36] "mage_7"  "mage_8"  "mage_9"  "mage_10" "mage_11" "mage_12"
head(toydata)
#>   id        x1        x2        x3 total age_1 age_2 age_3 age_4 age_5 age_6
#> 1  1 0.5108491 0.5202009 0.3148000  4132    NA    NA    NA    NA    NA    NA
#> 2  2 0.4646444 0.6100449 0.1927605  5443    NA    NA    NA    NA    NA    NA
#> 3  3 0.3253992 0.4715522 0.5539218  6355    NA    NA    NA    NA    NA    NA
#> 4  4 0.3963392 0.6752947 0.5494017  5647    NA    NA    NA    NA    NA    NA
#> 5  5 0.4253325 0.7066313 0.6955719  6065   803   765   795   760   565   764
#> 6  6 0.5800898 0.4553292 0.9862218  7331   910   908  1012   940   683   907
#>   age_7 age_8 age_9 age_10 age_11 age_12 fage_1 fage_2 fage_3 fage_4 fage_5
#> 1    NA    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#> 2    NA    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#> 3    NA    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#> 4    NA    NA    NA     NA     NA     NA     NA     NA     NA     NA     NA
#> 5   582   387   249    188    125     82    337    352    398    448    350
#> 6   683   467   323    257    150     91    428    381    405    470    348
#>   fage_6 fage_7 fage_8 fage_9 fage_10 fage_11 fage_12 mage_1 mage_2 mage_3
#> 1     NA     NA     NA     NA      NA      NA      NA     NA     NA     NA
#> 2     NA     NA     NA     NA      NA      NA      NA     NA     NA     NA
#> 3     NA     NA     NA     NA      NA      NA      NA     NA     NA     NA
#> 4     NA     NA     NA     NA      NA      NA      NA     NA     NA     NA
#> 5    329    285    174    115      94      66      48    466    413    397
#> 6    463    335    210    184     111      68      56    482    527    607
#>   mage_4 mage_5 mage_6 mage_7 mage_8 mage_9 mage_10 mage_11 mage_12
#> 1     NA     NA     NA     NA     NA     NA      NA      NA      NA
#> 2     NA     NA     NA     NA     NA     NA      NA      NA      NA
#> 3     NA     NA     NA     NA     NA     NA      NA      NA      NA
#> 4     NA     NA     NA     NA     NA     NA      NA      NA      NA
#> 5    312    215    435    297    213    134      94      59      34
#> 6    470    335    444    348    257    139     146      82      35